django_project_boilerplate-master/requirements.txt
asgiref==3.2.7
autopep8==1.4.4
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.8
django-allauth==0.42.0
django-countries==6.1.2
django-crispy-forms==1.9.1
django-filter==2.2.0
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.6.0
Markdown==3.2.2
oauthlib==3.1.0
pep8==1.7.1
Pillow==7.1.2
psycopg2==2.8.5
pycodestyle==2.5.0
python3-openid==3.1.0
pytz==2018.5
requests==2.23.0
requests-oauthlib==1.3.0
sqlparse==0.2.4
stripe==2.48.0
urllib3==1.25.9
zipp==3.1.0
